updates "countsize(a)" macro
* updates "countsize(a)" macro to incorporate static conversion to int

+++ lib/define.h
@@ -107,7 +107,7 @@
 #define FORI(val) for (int i=0; i < val; i++)
 #define FORJ(val) for (int j=0; j < val; j++)
 #define FORIJ(val1, val2) for (int i=0; i < val1; i++) for (int j=0; j < val2; j++)
-#define countSize(a)    (sizeof(a) / sizeof((a)[0]))
+#define countSize(a)    (  static_cast<int> (sizeof(a) / sizeof((a)[0])) )
 
 typedef float  float32_t;
 typedef double float64_t;
